+++ R/Tagged_table.R
@@ -133,12 +133,12 @@ tt_meta <- function(
   subtitle = NULL,
   footer = NULL
 ){
-  assert_that(rlang::is_scalar_atomic(table_id) || is.null(table_id))
-  assert_that(rlang::is_scalar_atomic(title) || is.null(title))
+  assert_that(is_scalar_atomic(table_id) || is.null(table_id))
+  assert_that(is_scalar_atomic(title) || is.null(title))
 
-  assert_that(is.null(longtitle) || rlang::is_atomic(longtitle))
-  assert_that(is.null(subtitle)  || rlang::is_atomic(subtitle))
-  assert_that(is.null(footer)    || rlang::is_atomic(footer))
+  assert_that(is.null(longtitle) || is.atomic(longtitle))
+  assert_that(is.null(subtitle)  || is.atomic(subtitle))
+  assert_that(is.null(footer)    || is.atomic(footer))
 
   if(all(
       is.null(table_id),
@@ -357,7 +357,7 @@ table_id <- function(x){
 #'   `list(longtitle = value)`
 #'
 assign_tt_meta <- function(x, assignment){
-  assert_that(rlang::is_scalar_list(assignment))
+  assert_that(is_scalar_list(assignment))
   assert_that(identical(
     length(names(assignment)), 1L
   ))